+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ### GV 1.56:
 
 ## Release 1.2.3
+- Fixed arenas for Thunderstore.
 - Audio clips that are too big will no longer be attempted to be cached (due to the 2GB limit of byte arrays).
+- Save file game version conversion hairstyle material fix.
 
 ## Release 1.2.2
 - Added experimental support for doors.